Nemolite2d manual C++ has two implementations: a raw pointers one and a std::vectors one
